refactor: rename constants file which contains BYTES_IN_KB to logger.ts
